+++ retrocache/src/main/java/com/andiag/retrocache/annotations/Caching.java
@@ -1,0 +1,18 @@
+package com.andiag.retrocache.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Created by Canalejas on 02/02/2017.
+ * Annotation used to avoid a {@link retrofit2.http.GET} method to be cached.
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface Caching {
+
+    boolean active() default true;
+
+}