+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# NEXT
+    - Added `@Caching(active = true)` annotation. Used to disable cache for some calls.
+
 # 1.1.1
     - Added `clone()` method to `CachedCall` interface to avoid extra casting.
     - Implemented new tests and fixed old ones.